fix bug of delete order

+++ service/src/main/java/cn/momia/service/web/ctrl/deal/OrderController.java
@@ -59,13 +59,19 @@ public class OrderController extends AbstractController {
         return skuService.unlock(order.getSkuId(), order.getCount());
     }
 
-    @RequestMapping(value = "/{id}", method = RequestMethod.POST)
+    @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     public ResponseMessage deleteOrder(@PathVariable long id, @RequestParam String utoken) {
         User user = userService.getByToken(utoken);
         if (!user.exists()) return new ResponseMessage(ErrorCode.FORBIDDEN, "user not login");
 
+        Order order = orderService.get(id);
+        if (!order.exists()) return new ResponseMessage(ErrorCode.FORBIDDEN, "order not exists");
+
         if (!orderService.delete(id, user.getId())) return new ResponseMessage(ErrorCode.FORBIDDEN, "fail to delete order");
-        return new ResponseMessage("delete order successfully");
+
+        if (!unlockSku(order)) LOGGER.error("fail to unlock sku, skuId: {}, count: {}", new Object[] { order.getSkuId(), order.getCount() });
+
+        return ResponseMessage.SUCCESS;
     }
 
     @RequestMapping(method = RequestMethod.GET)
